archive/2025-09-12/vibevoice_no_dup/requirements.txt
torch
transformers>=4.51.3
accelerate
soundfile
torchaudio
numpy
peft>=0.11.0
sentencepiece
pyyaml
scipy
librosa
tqdm
typing_extensions
datasets
speechbrain


